requirements.txt
fastapi
uvicorn[standard]
pytest
pytest-playwright
requests
